net: mdio: Add netlink interface
This adds a netlink interface to make reads/writes to mdio busses. This
makes it easier to debug devices. This is especially useful when there
is a PCS involved (and the ethtool reads are faked) or when there is no
MAC associated with a PHY.

This driver was written by Tobias Waldekranz <tobias@waldekranz.com>. It
is adapted from the version he released with mdio-tools.

[1] https://github.com/wkz/mdio-tools
Signed-off-by: Sean Anderson <sean.anderson@seco.com>