net: macb: Clean up macb_validate
As the number of interfaces grows, the number of if statements grows
ever more unweildy. Clean everything up a bit by using a switch
statement. This reduces the number of if statements, and the number of
times the same condition is checked. No functional change intended.

Signed-off-by: Sean Anderson <sean.anderson@seco.com>
Commit-notes:
This patch was originally submitted as [1].

There is another approach which could be used here. We could do
something like

macb_set_mode(bp, mask, iface)
{
	switch (iface) {
	case 10GBASER:
		if (...)
			phylink_set_10g_modes(mask);
		else
			return -EINVAL;
		break;
	case GMII:
	/* etc etc */
	}
	return 0;
}

macb_validate(...)
{
	if (state->interface == PHY_INTERFACE_MODE_NA) {
		macb_set_mode(bp, mask, PHY_INTERFACE_MODE_10GBASER);
		macb_set_mode(bp, mask, PHY_INTERFACE_MODE_GMII);
		macb_set_mode(bp, mask, PHY_INTERFACE_MODE_MII);
	} else if (macb_set_mode(bp, mask, state->interface)) {
		bitmap_zero(supported, __ETHTOOL_LINK_MODE_MASK_NBITS);
		return;
	}

	/* etc etc */
}

which has the advantage of much cleaner logic in the switch statement.

[1] https://lore.kernel.org/netdev/20211004191527.1610759-9-sean.anderson@seco.com/
END
Series-changes: 2
- Fix polarity of `one` being inverted
- Only set gigabit modes for NA if macb_is_gem()

Series-changes: 3
- Remove labels/gotos in favor of explicit zeroing. Hopefully this
  better illustrates the "linear" flow of the logic.
- Add comment with overview of the flow of the logic.

Series-changes: 4
- Move an if statement around