net: dsa: felix: Convert to use device-mode PCS
This converts the ocelot felix driver to use PCS devices, instead of
attaching the Lynx library to an MDIO device. Like ENETC, the PCSs are
children of the DSA driver, so device links provide no ordering.
However, we cannot prevent registration of the DSA driver, since we
only create the MDIO bus (and hence the PCSs) in setup(). Instead, we
just call dsa_switch_shutdown in unbind. This will bring down all DSA
ports, releasing their PCSs. I am not sot sure how this will behave if
we rebind the PCS drivers. Maybe we need a dsa_switch_startup which
calls dsa_tree_setup? At the very least we will avoid a segfault with
this method.

Signed-off-by: Sean Anderson <sean.anderson@seco.com>