net: enetc: Convert to use device-mode PCSs
This converts the ENETC driver to use PCS devices, instead of attaching
the Lynx library to an MDIO device. Because the PCS is likely a child of
the ethernet device, device links will not provide any ordering. To
avoid use-after-free bugs, we use a composite device to defer
phylink_create and netdev_register until after the PCS driver has been
bound. During normal operation, there will be no problem (since Linux
eagerly binds drivers if they are available, and -EPROBE_DEFER will be
returned if they are not). However, this is necessary to ensure that
we stop phylink before the PCS driver is unbound.

Signed-off-by: Sean Anderson <sean.anderson@seco.com>