dt-bindings: net: Add binding for Xilinx PCS
This adds a binding for the Xilinx 1G/2.5G Ethernet PCS/PMA or SGMII
LogiCORE IP. This device is a soft device typically used to adapt
between GMII and SGMII or 1000BASE-X (possbilty in combination with a
serdes). The "standard" property is roughly analogous to the interface
property of ethernet controllers, except that it has an additional value
used to indicate that dynamic switching is supported. Note that
switching is supported only between SGMII and 1000BASE-X, and only if
the appropriate parameter is set when the device is synthesized. The
property name was chosen to align with the terminology in the datasheet.
I also considered "mdi", but that is a bit of a misnomer in the case of
SGMII (which has another phy before the MDI), and 1000BASE-X isn't
really an MDI either.

Signed-off-by: Sean Anderson <sean.anderson@linux.dev>
Patch-cc: robh
Patch-cc: msimek
Patch-cc: devicetree@vger.kernel.org