net: pcs: Add Xilinx PCS driver
This adds support for the Xilinx 1G/2.5G Ethernet PCS/PMA or SGMII
device. This is a soft device which converts between GMII and either
SGMII, 1000Base-X, or 2500Base-X. If configured correctly, it can also
switch between SGMII and 1000BASE-X at runtime. Thoretically this is
also possible for 2500Base-X, but that requires reconfiguring the
serdes. The exact capabilities depend on synthesis parameters, so they
are read from the devicetree.

This device has a c22-compliant PHY interface, so for the most part we
can just use the phylink helpers. This device supports an interrupt
which is triggered on autonegotiation completion. I'm not sure how
useful this is, since we can never detect a link down (in the PCS).

This device supports sharing some logic between different
implementations of the device. In this case, one device contains the "shared
logic" and the clocks are connected to other devices. To coordinate
this, one device registers a clock that the other devices can request.
The clock is enabled in the probe function by releasing the device from
reset. There are no othe software controls, so the clock ops are empty.

Later in this series, we will convert the Xilinx AXI Ethernet driver to
use this PCS. To help out, we provide a compatibility function to bind
this driver in the event the MDIO device has no compatible.

Signed-off-by: Sean Anderson <sean.anderson@linux.dev>
Patch-cc: msimek
Patch-cc: Robert Hancock <robert.hancock@calian.com>

Series-prefix: net-next
Series-to: netdev@vger.kernel.org
Series-to: David S. Miller <davem@davemloft.net>
Series-to: Jakub Kicinski <kuba@kernel.org>
Series-to: linux-kernel@vger.kernel.org
Series-cc: Russell King <linux@armlinux.org.uk>
Series-cc: Heiner Kallweit <hkallweit1@gmail.com>
Series-cc: Andrew Lunn <andrew@lunn.ch

Cover-letter:
Add core PCS support
This series adds support for Xilinx PCS devices. Adding support for this
device has required some plumbing work related to PCSs in general, and I
would appreciate feedback in that area.
END